move BackgroundColors class to own file, add some default configs (easy to change for development)
